requirements.txt
anyio==2.0.2
APScheduler==3.6.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cifar10==1.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fpdf==1.7.2
func-timeout==4.3.5
geneticalgorithm==1.0.2
graphviz==0.17
idna==2.10
inflection==0.5.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-server==1.2.2
jupyterlab==3.0.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.2
jupyterthemes==0.20.0
kiwisolver==1.3.1
lesscpy==0.15.0
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
more-itertools==8.6.0
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
nltk==3.6.3
notebook==6.2.0
numpy==1.19.5
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.1.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.11
pycparser==2.20
pydotplus==2.0.2
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-telegram-bot==13.11
pytz==2020.5
pytz-deprecation-shim==0.1.0.post0
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==21.0.1
Quandl==3.6.0
regex==2021.9.24
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.2.1
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
